// pc/src/services/apis/models/ProjectFilingInfoVo.ts
export interface ProjectFilingInfoVo {
    /**
     * 主键
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    id?: string;
    /**
     * 市（区），如：靖江市、泰兴市等
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    district?: string;
    /**
     * park
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    park?: string;
    /**
     * 项目名称
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    projectName?: string;
    /**
     * 备案时的项目名称
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    filingProjectName?: string;
    /**
     * 申报单位
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    applicationUnit?: string;
    /**
     * 项目投资额 (万元)
     * @type {number}
     * @memberof ProjectFilingInfoVo
     */
    investmentAmount?: number;
    /**
     * 项目代码
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    projectCode?: string;
    /**
     * 建设规模
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    constructionScale?: string;
    /**
     * 备案证号
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    filingCertificateNo?: string;
    /**
     * 申请备案时间
     * @type {Date}
     * @memberof ProjectFilingInfoVo
     */
    applyFilingTime?: Date;
    /**
     * 完成备案时间
     * @type {Date}
     * @memberof ProjectFilingInfoVo
     */
    completeFilingTime?: Date;
    /**
     * 是否外资项目 (0:否, 1:是)
     * @type {boolean}
     * @memberof ProjectFilingInfoVo
     */
    isForeignInvestment?: boolean;
    /**
     * 法人单位信息
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    legalEntityInfo?: string;
    /**
     * 备案部门
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    filingDepartment?: string;
    /**
     * 项目类型（工业、服务业）
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    projectType?: string;
    /**
     * 产业方向（工业项目选“8+13+X”类别或“其他”）
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    industryDirection?: string;
    /**
     * 用地类型
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    landUseType?: string;
    /**
     * 供地进展（新增用地项目填写）
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    landSupplyProgress?: string;
    /**
     * 环评进展
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    environmentalAssessmentStatus?: string;
    /**
     * 安评情况
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    safetyAssessmentStatus?: string;
    /**
     * 能评情况
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    energyAssessmentStatus?: string;
    /**
     * 施工图审查情况
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    constructionDrawingReviewStatus?: string;
    /**
     * 施工许可情况
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    constructionPermitStatus?: string;
    /**
     * 规划许可情况
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    planningPermitStatus?: string;
    /**
     * 安评情况颜色
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    safetyAssessmentStatusColor?: string;
    /**
     * 能评情况颜色
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    energyAssessmentStatusColor?: string;
    /**
     * 施工许可情况颜色
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    constructionPermitStatusColor?: string;
    /**
     * 环评进展颜色
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    environmentalAssessmentStatusColor?: string;
    /**
     * 施工图审查情况颜色
     * @type {string}
     * @memberof ProjectFilingInfoVo
     */
    constructionDrawingReviewStatusColor?: string;
}